backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-dotenv==1.0.0

# AI & Memory (Critical Fix #1 & #2)
# letta is run as a separate server — not a pip dependency
google-generativeai>=0.4.0
chromadb>=0.4.22

# ML
scikit-learn>=1.3.0
xgboost>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
joblib>=1.3.0

# Graph & Causal Analysis
networkx>=3.1
dowhy>=0.11.0

# Utilities
httpx==0.27.0
aiofiles>=23.2.0
python-multipart>=0.0.6
google-api-python-client>=2.150.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.1

# Testing
pytest>=7.4.0

# Phase 1 & 4 Enhancements
alembic>=1.13.0
psycopg2-binary>=2.9.9
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
apscheduler>=3.10.4
websockets>=12.0
cryptography>=42.0.0
pytest-cov>=4.1.0
